@@ -4,32 +4,33 @@
   "private": true,
   "proxy": "http://localhost:5000",
   "dependencies": {
-    "@material-ui/core": "^4.9.10",
-    "@material-ui/icons": "^4.9.1",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "axios": "^0.19.2",
-    "jwt-decode": "^4.0.0",
-    "moment": "^2.27.0",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
+    "@emotion/react": "^11.13.3",
+    "@emotion/styled": "^11.13.0",
+    "@mui/icons-material": "^5.0.0",
+    "@mui/material": "^5.16.7",
+    "@mui/styles": "^6.0.2",
+    "@react-oauth/google": "^0.12.1",
+    "@reduxjs/toolkit": "^2.2.7",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.0.0",
+    "@testing-library/user-event": "^14.0.0",
+    "axios": "^0.27.2",
+    "jwt-decode": "^3.1.2",
+    "moment": "^2.29.3",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-file-base64": "^1.0.3",
-    "react-google-login": "^5.2.2",
-    "react-redux": "^7.1.3",
-    "react-router-dom": "^6.26.1",
-    "react-scripts": "3.4.1",
-    "redux": "^4.0.5",
-    "redux-thunk": "^2.3.0"
+    "react-redux": "^8.0.2",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.0",
+    "redux": "^5.0.1",
+    "redux-thunk": "^3.1.0"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
   },
   "browserslist": {
     "production": [
@@ -44,10 +45,6 @@
     ]
   },
   "devDependencies": {
-    "eslint-config-airbnb": "^18.2.0",
-    "eslint-plugin-import": "^2.22.0",
-    "eslint-plugin-jsx-a11y": "^6.3.1",
-    "eslint-plugin-react": "^7.20.6",
-    "eslint-plugin-react-hooks": "^4.1.0"
+    "@babel/plugin-proposal-private-property-in-object": "^7.18.6"
   }
 }
