@@ -1,21 +1,24 @@
 import React from 'react';
-import { Container } from '@material-ui/core';
+import { Container } from '@mui/material';
 import Navbar from './components/Navbar/Navbar';
 import Home from './components/Home/Home';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import Auth from './components/Auth/Auth';
+import { GoogleOAuthProvider } from '@react-oauth/google';
 
 const App = () => {
   return (
-    <BrowserRouter>
-      <Container maxWidth="lg">
-        <Navbar />
-        <Routes>
-          <Route path="/" element={<Home />} />
-          <Route path="/auth" element={<Auth />} />
-        </Routes>
-      </Container>
-    </BrowserRouter>
+    <GoogleOAuthProvider clientId="544061686466-0prcjj4rkb7b9gtkfmm8m7eta4m1shjt.apps.googleusercontent.com">
+      <BrowserRouter>
+        <Container maxWidth="lg">
+          <Navbar />
+          <Routes>
+            <Route path="/" element={<Home />} />
+            <Route path="/auth" element={<Auth />} />
+          </Routes>
+        </Container>
+      </BrowserRouter>
+    </GoogleOAuthProvider>
   );
 };
 
