@@ -1,14 +1,14 @@
 
-import { makeStyles } from '@material-ui/core/styles';
+import { makeStyles } from '@mui/styles';
 
-export default makeStyles((theme) => ({
+export default makeStyles(() => ({
   root: {
     '& .MuiTextField-root': {
-      margin: theme.spacing(1),
+      margin: '8px',
     },
   },
   paper: {
-    padding: theme.spacing(2),
+    padding: '16px',
   },
   form: {
     display: 'flex',
