@@ -1,7 +1,7 @@
-import { makeStyles } from '@material-ui/core/styles';
-import { deepPurple } from '@material-ui/core/colors';
+import { makeStyles } from '@mui/styles';
+import { deepPurple } from '@mui/material/colors';
 
-export default makeStyles((theme) => ({
+export default makeStyles(() => ({
   appBar: {
     borderRadius: 15,
     margin: '30px 0',
@@ -37,7 +37,7 @@ export default makeStyles((theme) => ({
     alignItems: 'center',
   },
   purple: {
-    color: theme.palette.getContrastText(deepPurple[500]),
+    color: '#fff', // Sử dụng màu trắng trực tiếp
     backgroundColor: deepPurple[500],
   },
 }));
