@@ -1,9 +1,103 @@
-import React from 'react'
+import React, { useState } from 'react';
+import { Avatar, Button, Paper, Grid, Typography, Container } from '@mui/material';
+import LockOutlinedIcon from '@mui/icons-material/LockOutlined';
+import { GoogleLogin } from '@react-oauth/google';
+import useStyles from './styles';
+import Input from './Input';
+import Icon from './icon';
+import jwtDecode from 'jwt-decode';
+import { useDispatch } from 'react-redux';
+import { useNavigate } from 'react-router-dom';
+import { signin, signup } from '../../actions/auth';
+
+const intialState = { firstName: '', lastName: '', email: '', password: '', confirmPassword: '' }; 
 
 const Auth = () => {
-  return (
-    <div>Auth</div>
-  )
-}
+  const classes = useStyles();
+  const [showPassword, setShowPassword] = useState(false);
+  const [isSignup, setIsSignup] = useState(false);
+  const dispatch = useDispatch();
+  const handleShowPassword = () => setShowPassword((prevShowPassword) => !prevShowPassword);
+  const navigate = useNavigate();
 
-export default Auth
+  const [formData, setFormData] = useState(intialState);
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    if (isSignup) {
+      dispatch(signup(formData, navigate));
+    } else {
+      dispatch(signin(formData, navigate));
+    }
+  };
+
+  const handleChange = (e) => {
+    setFormData({ ...formData, [e.target.name]: e.target.value });
+  };
+
+  const switchMode = () => {
+    setIsSignup((prevIsSignup) => !prevIsSignup);
+    setShowPassword(false);
+  };
+
+  return (
+    <Container component="main" maxWidth="xs">
+      <Paper className={classes.paper} elevation={3}>
+        <Avatar className={classes.avatar}>
+          <LockOutlinedIcon />
+        </Avatar>
+        <Typography variant="h5">{isSignup ? 'Sign Up' : 'Sign In'}</Typography>
+        <form className={classes.form} onSubmit={handleSubmit}>
+          <Grid container spacing={2}>
+            {isSignup && (
+              <>
+                <Input name="firstName" label="First Name" handleChange={handleChange} autoFocus half />
+                <Input name="lastName" label="Last Name" handleChange={handleChange} half />
+              </>
+            )}
+            <Input name="email" label="Email Address" handleChange={handleChange} type="email" />
+            <Input name="password" label="Password" handleChange={handleChange} type={showPassword ? 'text' : 'password'} handleShowPassword={handleShowPassword} />
+            {isSignup && <Input name="confirmPassword" label="Repeat Password" handleChange={handleChange} type="password" />}
+          </Grid>
+          <Button type="submit" fullWidth variant="contained" color="primary" className={classes.submit}>
+            {isSignup ? 'Sign Up' : 'Sign In'}
+          </Button>
+          <GoogleLogin
+              onSuccess={credentialResponse => {
+                const decoded = jwtDecode(credentialResponse.credential);
+                dispatch({ type: 'AUTH', data: { result: decoded, token: credentialResponse.credential } });
+
+                navigate('/');
+                console.log(decoded);
+              }}
+              onError={() => {
+                console.log('Login Failed');
+              }}
+              useOneTap
+              render={(renderProps) => (
+                <Button
+                  className={classes.googleButton}
+                  color="primary"
+                  fullWidth
+                  onClick={renderProps.onClick}
+                  disabled={renderProps.disabled}
+                  startIcon={<Icon />}
+                  variant="contained"
+                >
+                  Google Sign In
+                </Button>
+              )}
+            />
+          <Grid container justifyContent="flex-end">
+            <Grid item>
+              <Button onClick={switchMode}>
+                {isSignup ? 'Already have an account? Sign in' : "Don't have an account? Sign Up"}
+              </Button>
+            </Grid>
+          </Grid>
+        </form>
+      </Paper>
+    </Container>
+  );
+};
+
+export default Auth;
