@@ -1,4 +1,5 @@
 import { combineReducers } from "redux";
 import posts from "./posts";
+import auth from "./auth";
 
-export const reducers = combineReducers({ posts });
+export const reducers = combineReducers({ posts, auth });
