@@ -3,19 +3,20 @@ import mongoose from 'mongoose';
 const postSchema = mongoose.Schema({
     title: String,
     message: String,
+    name : String,
     creator: String,
     tags : [String],
     selectedFile: String,
-    likeCount: {
-        type: Number,
-        default: 0
+    likes: {
+        type: [String],
+        default: []
     },
-    createdAt: {
+    createdAt: {    
         type: Date,
         default: new Date()
     }
 });
 
-const PostMessage = mongoose.model('PostMessage', postSchema);
+const PostMessage = mongoose.model('Post', postSchema);
 
 export default PostMessage;
