@@ -1,15 +1,17 @@
 import express from 'express';
 import { getPosts, createPosts, updatePost, deletePost, likePost } from '../controllers/posts.js';
 
+import auth from '../middleware/auth.js';
+
 const router = express.Router();
 
 router.get('/', getPosts);
 
-router.post('/', createPosts);
+router.post('/', auth, createPosts);
 
-router.patch('/:id', updatePost);
+router.patch('/:id', auth, updatePost);
 
-router.delete('/:id', deletePost);
+router.delete('/:id', auth, deletePost);
 
-router.patch('/:id/likePost', likePost);
+router.patch('/:id/likePost', auth, likePost);
 export default router;
